--- src/main/java/or/sopt/houme/domain/house/dto/request/IsLikeRequest.java
@@ -1,6 +1,0 @@
-package or.sopt.houme.domain.house.dto.request;
-
-public record IsLikeRequest(
-        boolean isLike
-) {
-}
