--- src/main/java/or/sopt/houme/domain/house/entity/mapping/HouseTaste.java
@@ -1,26 +1,0 @@
-package or.sopt.houme.domain.house.entity.mapping;
-
-import jakarta.persistence.*;
-import lombok.*;
-import or.sopt.houme.domain.house.entity.House;
-import or.sopt.houme.domain.taste.entity.Taste;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@AllArgsConstructor
-@Builder
-@Table(name = "house_tastes")
-public class HouseTaste {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "taste_id")
-    private Taste taste;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "house_id")
-    private House house;
-}
