+++ src/main/java/or/sopt/houme/domain/house/model/entity/InvalidHouseRequest.java
@@ -1,11 +1,11 @@
-package or.sopt.houme.domain.house.entity;
+package or.sopt.houme.domain.house.model.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import or.sopt.houme.domain.house.entity.enums.Equilibrium;
-import or.sopt.houme.domain.house.entity.enums.Form;
-import or.sopt.houme.domain.house.entity.enums.Structure;
-import or.sopt.houme.domain.user.entity.User;
+import or.sopt.houme.domain.house.model.entity.enums.Equilibrium;
+import or.sopt.houme.domain.house.model.entity.enums.Form;
+import or.sopt.houme.domain.house.model.entity.enums.Structure;
+import or.sopt.houme.domain.user.model.entity.User;
 import or.sopt.houme.global.entity.BaseEntity;
 
 @Entity
