+++ src/main/java/or/sopt/houme/domain/house/presentation/dto/HouseOptionDTO.java
@@ -1,4 +1,4 @@
-package or.sopt.houme.domain.house.dto;
+package or.sopt.houme.domain.house.presentation.dto;
 
 public record HouseOptionDTO(
         String code,
