+++ src/main/java/or/sopt/houme/domain/house/presentation/dto/request/IsLikeRequest.java
@@ -1,0 +1,6 @@
+package or.sopt.houme.domain.house.presentation.dto.request;
+
+public record IsLikeRequest(
+        boolean isLike
+) {
+}
