+++ src/main/java/or/sopt/houme/domain/house/repository/HouseRepository.java
@@ -1,6 +1,6 @@
 package or.sopt.houme.domain.house.repository;
 
-import or.sopt.houme.domain.house.entity.House;
+import or.sopt.houme.domain.house.model.entity.House;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
