+++ src/main/java/or/sopt/houme/domain/house/repository/HouseTasteRepository.java
@@ -1,7 +1,7 @@
 package or.sopt.houme.domain.house.repository;
 
-import or.sopt.houme.domain.house.entity.mapping.HouseTaste;
-import or.sopt.houme.domain.taste.entity.Taste;
+import or.sopt.houme.domain.house.model.entity.mapping.HouseTaste;
+import or.sopt.houme.domain.house.model.taste.entity.Taste;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
