+++ src/main/java/or/sopt/houme/domain/house/service/HouseServiceImpl.java
@@ -2,28 +2,28 @@ package or.sopt.houme.domain.house.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import or.sopt.houme.domain.floorPlan.entity.FloorPlan;
-import or.sopt.houme.domain.floorPlan.repository.FloorPlanRepository;
-import or.sopt.houme.domain.furniture.entity.Furniture;
+import or.sopt.houme.domain.house.model.floorPlan.entity.FloorPlan;
+import or.sopt.houme.domain.house.repository.floorPlan.FloorPlanRepository;
+import or.sopt.houme.domain.furniture.model.entity.Furniture;
 import or.sopt.houme.domain.furniture.repository.FurnitureRepository;
-import or.sopt.houme.domain.house.dto.HouseOptionDTO;
-import or.sopt.houme.domain.house.dto.LatestHouseConditionDTO;
-import or.sopt.houme.domain.house.dto.request.HouseSelectRequest;
-import or.sopt.houme.domain.house.dto.response.HouseIdResponse;
-import or.sopt.houme.domain.house.dto.response.HouseOptionsResponse;
-import or.sopt.houme.domain.house.entity.House;
-import or.sopt.houme.domain.house.entity.InvalidHouseRequest;
-import or.sopt.houme.domain.house.entity.enums.Activity;
-import or.sopt.houme.domain.house.entity.enums.Equilibrium;
-import or.sopt.houme.domain.house.entity.enums.Form;
-import or.sopt.houme.domain.house.entity.enums.Structure;
-import or.sopt.houme.domain.house.entity.mapping.HouseFloorPlan;
-import or.sopt.houme.domain.house.entity.mapping.HouseFurniture;
-import or.sopt.houme.domain.house.entity.mapping.HouseTaste;
+import or.sopt.houme.domain.house.presentation.dto.HouseOptionDTO;
+import or.sopt.houme.domain.house.presentation.dto.LatestHouseConditionDTO;
+import or.sopt.houme.domain.house.presentation.dto.request.HouseSelectRequest;
+import or.sopt.houme.domain.house.presentation.dto.response.HouseIdResponse;
+import or.sopt.houme.domain.house.presentation.dto.response.HouseOptionsResponse;
+import or.sopt.houme.domain.house.model.entity.House;
+import or.sopt.houme.domain.house.model.entity.InvalidHouseRequest;
+import or.sopt.houme.domain.house.model.entity.enums.Activity;
+import or.sopt.houme.domain.house.model.entity.enums.Equilibrium;
+import or.sopt.houme.domain.house.model.entity.enums.Form;
+import or.sopt.houme.domain.house.model.entity.enums.Structure;
+import or.sopt.houme.domain.house.model.entity.mapping.HouseFloorPlan;
+import or.sopt.houme.domain.house.model.entity.mapping.HouseFurniture;
+import or.sopt.houme.domain.house.model.entity.mapping.HouseTaste;
 import or.sopt.houme.domain.house.repository.*;
-import or.sopt.houme.domain.taste.entity.Taste;
-import or.sopt.houme.domain.taste.repository.taste.TasteRepository;
-import or.sopt.houme.domain.user.entity.User;
+import or.sopt.houme.domain.house.model.taste.entity.Taste;
+import or.sopt.houme.domain.house.repository.taste.taste.TasteRepository;
+import or.sopt.houme.domain.user.model.entity.User;
 import or.sopt.houme.global.api.ErrorCode;
 import or.sopt.houme.global.api.GeneralException;
 import or.sopt.houme.global.api.handler.HouseException;
